% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chk_col.R
\name{chk_col}
\alias{chk_col}
\title{Order the columns if needed and rename them}
\usage{
chk_col(df)
}
\arguments{
\item{df}{dataframe to check}
}
\description{
This function check the column order and rename them
}

