% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chk_filename.R
\name{chk_filename}
\alias{chk_filename}
\title{Check m70 filenames}
\usage{
chk_filename(dirpath)
}
\arguments{
\item{dirpath}{Path to the directory to treat}
}
\description{
This function check if m70 files are standard.
Files are considered standard if they begin with a standard timestamp (AAAA-MM-JJ HH_MM). If not the function try to correct it.
}
\details{
(Example : 2013-02-11 7_45.m70 become : 2013-02-11 07_45.m70 )
}

