% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extrcFlux.R
\name{extrcFlux}
\alias{extrcFlux}
\title{Calcul flux from measurements.}
\usage{
extrcFlux(df)
}
\arguments{
\item{path}{path to the file to clean}
}
\description{
This function compute flux from measurements
}

