% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNF}
\alias{getNF}
\title{Calculate a Net flux}
\usage{
getNF(RF = 0.1, Dchb_mm = 300, Hchb_mm = 300, Patm = 101300,
  T_Cel = 25)
}
\arguments{
\item{RF}{Raw Flux in ppm/s}

\item{Dchb_mm}{Chamber diameter in millimeter}

\item{Hchb_mm}{Chamber height in millimeter}

\item{Patm}{Atmospheric pressure in Pa}

\item{T_Cel}{Temperature inside the chamber}
}
\description{
This function calculate a net flux with chamber parameter
}

