% Generated by roxygen2 (4.0.1): do not edit by hand
\name{m70setup}
\alias{m70setup}
\title{Data preparation}
\usage{
m70setup(path)
}
\arguments{
\item{path}{Path to the directory to treat}
}
\description{
Input a path to a directory with m70 files. The function will prepare the data
for visualisation. By defaut m70 files will be converted in csv and written in
an CO2_csv directory, a csv file with all the data will be written in an "output"
directory
}

