% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/m70strsplit.R
\name{m70strsplit}
\alias{m70strsplit}
\title{Stringr wrapper to easily return parts of the m70 filenames}
\usage{
m70strsplit(filename, output = "name")
}
\arguments{
\item{filename}{filename to treat}

\item{output}{What part to return ?
"name" : return filename without extension
"date" : return date
"time" : return time}
}
\description{
This function allow to return information (date, time) from a m70 filename
}
\examples{
file <- "2013-10-01 08_03.m70"
splt_m70(file, "name")
splt_m70(file, "date")
splt_m70(file, "time")
}

