% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/m70tocsv.R
\name{m70tocsv}
\alias{m70tocsv}
\title{Convert \emph{Vaisala} files ".M70" to ".csv"}
\usage{
m70tocsv(path, newdir = T, dirname = "CO2_csv")
}
\arguments{
\item{path}{path to the directory to treat}

\item{newdir}{if TRUE the file are saved in a new directory}

\item{dirname}{a name for the output directory}
}
\description{
Convert all m70 from a directory to csv files
}

