% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pdfdplot.R
\name{pdfdplot}
\alias{pdfdplot}
\title{Set up diagnostic plots}
\usage{
pdfdplot(df, outname = "diagPlot.pdf", outpath = NULL)
}
\arguments{
\item{df}{data to proceed}

\item{outname}{a name for the output pdf file,
by default the name is "diagPlot.pdf"}

\item{outpath}{a path to a directory to save the file
by default a "output" directory is created in the working directory to store
the file}
}
\description{
This function print a diagnostic plot for multiple measurements
A fileid column is necessary
}

