% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regsetup}
\alias{regsetup}
\title{Select data used for regression}
\usage{
regsetup(df, path = getwd(), file = "regselection.csv")
}
\arguments{
\item{df}{data to proceed}

\item{path}{path to the setup file (working directory by defaut)}

\item{file}{setup file name ("regselection.csv" by defaut)}
}
\description{
This function add a keep column to the dataframe from an external file to
allow easy and reproducible changes in regression setup.
The external file is generated if not found.
The fileid column must contain the measurement ID to treat
The start column allow to put aside the desired number of starting points (interger)
The end column allow to put aside the desired number of ending points (interger)
The other column allow to put aside specifique point,
by giving a line number (interger)
The state column do nothing for now
}

