% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chk_misc.R
\name{substrRight}
\alias{substrRight}
\title{Misc}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{string to proceed}

\item{n}{starting on the right, how many characters to keep}
}
\description{
Misc
}

