% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timestp.R
\name{timestp}
\alias{timestp}
\title{Add a timestamp column based on date and time ones}
\usage{
timestp(df, replace = T)
}
\arguments{
\item{df}{data to proceed}

\item{replace}{if true after the timestamp column addition the date and time columns will be removed}
}
\description{
This function add a timestamp column based on the date and time columns
}

