% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/m70concat.R
\name{m70concat}
\alias{m70concat}
\title{Concat files}
\usage{
m70concat(path, id = "name", fileFormat = "m70")
}
\arguments{
\item{path}{Path to the files to concatenate}

\item{id}{defaut is "name" and add a file id (the actual name of the file)}

\item{fileFormat}{choose the input format of the files "m70" or "csv"}
}
\description{
This function Concat CO2 files either csv or m70
}

